+++ contracts/libraries/Tick.sol
@@ -8,18 +8,18 @@ library Tick {
   using PremiumPosition for mapping(address => PremiumPosition.Info);
 
   function addOwner(
-    mapping(uint24 => address[]) storage self,
+    mapping(uint32 => address[]) storage self,
     address owner,
-    uint24 tick
+    uint32 tick
   ) internal returns (uint256) {
     self[tick].push(owner);
     return self[tick].length - 1;
   }
 
   function removeOwner(
-    mapping(uint24 => address[]) storage self,
+    mapping(uint32 => address[]) storage self,
     uint256 ownerIndexToRemove,
-    uint24 tick
+    uint32 tick
   ) internal {
     address[] storage owners = self[tick];
     owners[ownerIndexToRemove] = owners[owners.length - 1];
@@ -27,30 +27,30 @@ library Tick {
   }
 
   function getLastOwnerInTick(
-    mapping(uint24 => address[]) storage self,
-    uint24 tick
+    mapping(uint32 => address[]) storage self,
+    uint32 tick
   ) internal view returns (address) {
     address[] memory owners = self[tick];
     return owners[owners.length - 1];
   }
 
   function getOwnerNumber(
-    mapping(uint24 => address[]) storage self,
-    uint24 tick
+    mapping(uint32 => address[]) storage self,
+    uint32 tick
   ) internal view returns (uint256) {
     return self[tick].length;
   }
 
-  function clear(mapping(uint24 => address[]) storage self, uint24 tick)
+  function clear(mapping(uint32 => address[]) storage self, uint32 tick)
     internal
   {
     delete self[tick];
   }
 
   function cross(
-    mapping(uint24 => address[]) storage self,
+    mapping(uint32 => address[]) storage self,
     mapping(address => PremiumPosition.Info) storage positions,
-    uint24 tick
+    uint32 tick
   ) internal view returns (uint256 policiesToRemove, uint256 capitalToRemove) {
     address[] memory owners = self[tick];
     for (uint256 i = 0; i < owners.length; i++) {
